Add: src/apis/gameManagement.jsをTypeScript対応させた
